Implement basic tests for RepositoryViewContext controllers
This implements most of the basic parts of the mentioned controller tests.
The tests currently work in isolation (such as `mvn clean test -D test=RepositoryViewContextControllerTest`) but not as a whole (such as `mvn clean test`).

There are concerns with the AOP interfering with the tests.

The restdocs (`.andDo()`) code is not implemented.