Fix broken tests, including replacing Junit 4 with Junit 5
Many of the problems get resolved when more thoroughly switching to Junit 5.

Moving ProceedingJoinPoint mock onto the class as an @MockBean.
This works better than calling `mock()` and does not result in NULL errors.

Some of the NULL errors on the RepositoryViewContextTransactionControllerTest tests were caused by returning TransactionDetails.
Instead, FedoraTransactionDetails should be returned when mocking.