Redesign context triple's property and metadata structure
The existing design requires a lot of extra work that with a valid redisign becomes unecessary.

Redesign the data structure for the property and metadata arrays for a given triple as a matrix.
Each matrix is an array keyed by the predicate of a given triple with the triples using that predicate are stored within the array at that key.

This alleviates the need for performing filtering on the predicate for every value.
This avoids the need to generate a temporary filtered array list.

Some of the existing design essentially mimics the behavior of how a matrix performs, so most of the changes involve removal of code.
The controller requires an added inner loop.