Defect D-01255: Metadata on scope not updating when add/removed until page refresh
This turns out to be two separate issues with identical symptoms.
1) Metadata is not being updated due to the count objects not being re-calculated on change.
2) It seems that the broadcast message is actually being sent but it is not being processed.

This addresses the first issue and provides a work-around for the second issue.
A follow-up commit is needed to properly resolve the second issue.

The structure of the count objects do not delete anything to ensure the scope is never lost for the counts.
A refresh action is needed to ensure that once a change is made, re-calculate all count objects.
The repository view section directive must also handle this refresh behavior.
An additional structure is added so that the totals for a given namespace predicate set can be processed without any additional loops.

By utilizing the $watch on the list object in the repository view section directive, the counts will effectively refreshed as if the broadcast was directly processed.
A more elegant solution is desired and may follow this commit.