Simplify controller test writing for common checks: controller exists and controller method exists
The preferred method would be to use a utility function that accepts a few parameters.
The problem with the preferred method is the underlying error reporting in karma does not show the full stack trace using this approach.

As a compromize, this implements the same process for every controller.
This still results in a size reduction of code by placing the differences in a single array.

The error reporting messages are also improved.
- Before: "Is the controller defined should be defined for admin"
- After: "Is the controller defined for ROLE_ADMIN"
- Before: "Are the scope methods defined update role should be defined"
- After: "Is the scope method updateRole defined"